+++ CHANGES.md
@@ -3,6 +3,7 @@
   * Add mach.charset middleware for automatically setting the response charset
   * Add autoIndex option to mach.file to automatically generate a directory listing
     when a directory is targeted in the request
+  * Add setters to Location properties
   * Renamed mach.forward => mach.proxy
   * Use mime.types from nginx instead of mime module
   * Fixed a bug in conn.json response helper